changed the size of the index to device, endpoint and netvar from 8 bit to 16. it is necessary at least for the netvar because there are endpoints (e.g., the lower arms) which contains more than 256 variables.
i had to change: EOtreenode, EOmatrix3d, EOnvsCfg, and EOrop.

svn path=/trunk/iCub/firmware/emBODY/arch-armcmx/body/embobj/embobj/; revision=17309

+++ emBODY/eBcode/arch-arm/embobj/embobj-api/EOnvsCfg.h
@@ -107,9 +107,9 @@ extern EOnvsCfg* eo_nvscfg_New(uint16_t ndevices, EOVstorageDerived* stg);
 extern eOresult_t eo_nvscfg_PushBackDevice(EOnvsCfg* p, eOnvscfgOwnership_t ownership, eOipv4addr_t ipaddress, eOuint16_fp_uint16_t hashfn_ep2index, uint16_t nendpoints);
 
 
-extern uint8_t eo_nvscfg_GetIndexOfLocalDevice(EOnvsCfg* p);
+extern uint16_t eo_nvscfg_GetIndexOfLocalDevice(EOnvsCfg* p);
 
-extern eOresult_t eo_nvscfg_ondevice_PushBackEndpoint(EOnvsCfg* p, uint8_t ondevindex, eOnvEP_t endpoint, eOuint16_fp_uint16_t hashfn_id2index, const EOconstvector* treeofnvs_con, const EOconstvector* datanvs_usr, uint32_t datanvs_size, eOvoid_fp_uint16_voidp_voidp_t datanvs_init, EOVmutexDerived* mtx);
+extern eOresult_t eo_nvscfg_ondevice_PushBackEndpoint(EOnvsCfg* p, uint16_t ondevindex, eOnvEP_t endpoint, eOuint16_fp_uint16_t hashfn_id2index, const EOconstvector* treeofnvs_con, const EOconstvector* datanvs_usr, uint32_t datanvs_size, eOvoid_fp_uint16_voidp_voidp_t datanvs_init, EOVmutexDerived* mtx);
 
 
 extern eOresult_t eo_nvscfg_data_Initialise(EOnvsCfg* p);
@@ -119,14 +119,14 @@ extern eOresult_t eo_nvscfg_data_Initialise(EOnvsCfg* p);
 
 
 extern eOresult_t eo_nvscfg_GetIndices(EOnvsCfg* p, 
-                                    eOipv4addr_t ip, eOnvEP_t ep, eOnvID_t id, 
-                                    uint8_t *ipindex, uint8_t *epindex, uint8_t *idindex); 
+                                        eOipv4addr_t ip, eOnvEP_t ep, eOnvID_t id, 
+                                        uint16_t *ipindex, uint16_t *epindex, uint16_t *idindex); 
 
 
-extern EOtreenode* eo_nvscfg_GetTreeNode(EOnvsCfg* p, uint8_t ondevindex, uint8_t onendpointindex, uint8_t onidindex);
+extern EOtreenode* eo_nvscfg_GetTreeNode(EOnvsCfg* p, uint16_t ondevindex, uint16_t onendpointindex, uint16_t onidindex);
 
 
-extern EOnv* eo_nvscfg_GetNV(EOnvsCfg* p, uint8_t ondevindex, uint8_t onendpointindex, uint8_t onidindex, EOtreenode* treenode, EOnv* nvtarget);
+extern EOnv* eo_nvscfg_GetNV(EOnvsCfg* p, uint16_t ondevindex, uint16_t onendpointindex, uint16_t onidindex, EOtreenode* treenode, EOnv* nvtarget);
